Use cupertino text field & button for onboarding
