+++ lib/views/onboarding/onboarding_view.dart
@@ -1,4 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:storypad/app_theme.dart';
 import 'package:storypad/gen/assets.gen.dart';
 import 'package:storypad/views/languages/languages_view.dart';
 import 'package:storypad/widgets/sp_fade_in.dart';
@@ -9,6 +11,9 @@ import 'package:storypad/widgets/base_view/base_route.dart';
 import 'onboarding_view_model.dart';
 
 part 'onboarding_content.dart';
+
+part 'local_widgets/next_button.dart';
+part 'local_widgets/nickname_field.dart';
 
 class OnboardingRoute extends BaseRoute {
   OnboardingRoute();
