+++ lib/views/onboarding/steps/step_1/onboarding_step_1_content.dart
@@ -45,9 +45,17 @@ class _OnboardingStep1Content extends StatelessWidget {
   }
 
   Widget buildActionButton(BuildContext context) {
-    return OutlinedButton(
-      child: Text(tr("button.next")),
-      onPressed: () => viewModel.next(context),
-    );
+    if (AppTheme.isCupertino(context)) {
+      return CupertinoButton.filled(
+        sizeStyle: CupertinoButtonSize.small,
+        onPressed: () => viewModel.next(context),
+        child: Text(tr("button.next")),
+      );
+    } else {
+      return OutlinedButton(
+        child: Text(tr("button.next")),
+        onPressed: () => viewModel.next(context),
+      );
+    }
   }
 }
