+++ lib/views/onboarding/steps/step_2/onboarding_step_2_content.dart
@@ -90,9 +90,17 @@ class _OnboardingStep2Content extends StatelessWidget {
   }
 
   Widget buildActionButton(BuildContext context) {
-    return OutlinedButton(
-      child: Text(tr("button.next")),
-      onPressed: () => viewModel.next(context),
-    );
+    if (AppTheme.isCupertino(context)) {
+      return CupertinoButton.filled(
+        sizeStyle: CupertinoButtonSize.small,
+        onPressed: () => viewModel.next(context),
+        child: Text(tr("button.next")),
+      );
+    } else {
+      return OutlinedButton(
+        child: Text(tr("button.next")),
+        onPressed: () => viewModel.next(context),
+      );
+    }
   }
 }
