+++ lib/views/onboarding/steps/step_2/onboarding_step_2_view.dart
@@ -1,6 +1,8 @@
 import 'dart:ui';
 
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:storypad/app_theme.dart';
 import 'package:storypad/gen/assets.gen.dart';
 import 'package:storypad/views/onboarding/local_widgets/click_animation.dart';
 import 'package:storypad/views/onboarding/local_widgets/fade_in_builder.dart';
