+++ lib/views/onboarding/steps/step_3/onboarding_step_3_view.dart
@@ -1,6 +1,8 @@
 import 'dart:ui';
 
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:storypad/app_theme.dart';
 import 'package:storypad/views/onboarding/local_widgets/click_animation.dart';
 import 'package:storypad/views/onboarding/local_widgets/end_drawer_screenshot.dart';
 import 'package:storypad/views/onboarding/local_widgets/fade_in_builder.dart';
