+++ lib/widgets/bottom_sheets/sp_nickname_bottom_sheet.dart
@@ -76,7 +76,7 @@ class SpNicknameBottomSheet extends BaseBottomSheet {
 
           if (AppTheme.isCupertino(context)) {
             return CupertinoButton.filled(
-              sizeStyle: CupertinoButtonSize.medium,
+              sizeStyle: CupertinoButtonSize.small,
               onPressed: unchanged ? null : () => save(context, controller),
               child: nickname == null ? Text(tr("button.save")) : Text(tr("button.update")),
             );
